[create] 表示用コンポーネントを作成
